Fix data importer and extractor
This commit fixes the issues with the data importer
such as formating of the data to sync with the input
which is acceptable by dynamodb.

In the data importer we are now parsing the string
as json object which is the error we were getting
previously.

I also wrote some independent tests which can help
debug issues if-any in future.

create-table just creates a basic dynamodb table.